Move signInAnd inside FirebaseAuthService, remove unneeded FirebaseDbService implementation
