+++ app/src/main/java/net/squanchy/schedule/ScheduleModule.java
@@ -4,7 +4,6 @@ import android.content.Context;
 
 import net.squanchy.navigation.Navigator;
 import net.squanchy.service.firebase.FirebaseDbService;
-import net.squanchy.service.firebase.injection.DbServiceType;
 import net.squanchy.service.repository.EventRepository;
 
 import dagger.Module;
@@ -20,7 +19,7 @@ class ScheduleModule {
     }
 
     @Provides
-    ScheduleService scheduleService(@DbServiceType(DbServiceType.Type.AUTHENTICATED) FirebaseDbService dbService, EventRepository eventRepository) {
+    ScheduleService scheduleService(FirebaseDbService dbService, EventRepository eventRepository) {
         return new ScheduleService(dbService, eventRepository);
     }
