+++ app/src/main/java/net/squanchy/service/firebase/injection/FirebaseModule.java
@@ -5,10 +5,9 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
 import net.squanchy.injection.ApplicationLifecycle;
-import net.squanchy.service.firebase.AuthenticatedFirebaseDbService;
 import net.squanchy.service.firebase.FirebaseAuthService;
 import net.squanchy.service.firebase.FirebaseDbService;
-import net.squanchy.service.firebase.UnauthenticatedFirebaseDbService;
+import net.squanchy.service.firebase.AuthenticatedFirebaseDbService;
 
 import dagger.Module;
 import dagger.Provides;
@@ -33,18 +32,10 @@ public class FirebaseModule {
 
     @ApplicationLifecycle
     @Provides
-    @DbServiceType(DbServiceType.Type.AUTHENTICATED)
-    FirebaseDbService authenticatedDbService(
-            @DbServiceType(DbServiceType.Type.UNAUTHENTICATED) FirebaseDbService dbService,
+    FirebaseDbService firebaseDbService(
+            DatabaseReference database,
             FirebaseAuthService authService
     ) {
-        return new AuthenticatedFirebaseDbService(dbService, authService);
-    }
-
-    @ApplicationLifecycle
-    @Provides
-    @DbServiceType(DbServiceType.Type.UNAUTHENTICATED)
-    FirebaseDbService unauthenticatedDbService(DatabaseReference database) {
-        return new UnauthenticatedFirebaseDbService(database);
+        return new AuthenticatedFirebaseDbService(database, authService);
     }
 }
