+++ app/src/main/java/net/squanchy/service/repository/injection/RepositoryModule.java
@@ -1,7 +1,6 @@
 package net.squanchy.service.repository.injection;
 
 import net.squanchy.service.firebase.FirebaseDbService;
-import net.squanchy.service.firebase.injection.DbServiceType;
 import net.squanchy.service.repository.EventRepository;
 import net.squanchy.service.repository.SpeakerRepository;
 import net.squanchy.support.lang.Checksum;
@@ -14,7 +13,7 @@ public class RepositoryModule {
 
     @Provides
     EventRepository eventService(
-            @DbServiceType(DbServiceType.Type.AUTHENTICATED) FirebaseDbService dbService,
+            FirebaseDbService dbService,
             Checksum checksum,
             SpeakerRepository speakerRepository
     ) {
@@ -23,7 +22,7 @@ public class RepositoryModule {
 
     @Provides
     SpeakerRepository speakerRepository(
-            @DbServiceType(DbServiceType.Type.AUTHENTICATED) FirebaseDbService dbService,
+            FirebaseDbService dbService,
             Checksum checksum
     ) {
         return new SpeakerRepository(dbService, checksum);
